Hacking rule to forbid resource unsafe fixtures
Extend the existing T105, which was missing checks for
tearDownClass anyways. Forbid overriding setUpClass
and tearDownClass except for tempest/test.py where the
test base class is defined.

To be able to enforce the rule, fixing setUpClass
which was added with new tests before this patch could
merge.

Partially-implements bp resource-cleanup

Change-Id: Ib4d98c2ff8776ea1379a044b5a30fb02e351ce75